fix: address 26 issues from full codebase review
Security:
- Remove plaintext SharedPreferences fallback in CredentialManager
- Gate HTTP BODY logging behind BuildConfig.DEBUG across all OkHttp clients
- Enable isMinifyEnabled and isShrinkResources for release builds
- Set allowBackup=false to prevent ADB backup of credential files
- Add base-config to network_security_config.xml for API 26-27
- Restrict AuthService.storedPassword visibility to internal

Build & Config:
- Add signingConfigs placeholder for release signing
- Enable BuildConfig generation (buildConfig=true for AGP 9.x)
- Add Room schema export directory (exportSchema=true)
- Fix double-slash in CourseService API URL (api//courses -> api/courses)

Runtime Bugs:
- Replace mutableMapOf with ConcurrentHashMap in NtustSessionManager cookie store
- Fix ClassTableViewModel._isLoading stuck true on missing credentials
- Preserve assignment isCompleted state across remote syncs
- Add SupervisorJob to AppState CoroutineScope
- Fix MoodleApiModels.events to default to emptyList() for Gson null safety
- Change Moodle timestamp fields from Int to Long to prevent 2038 overflow
- Invalidate session when cookie store is empty after process death
- Wrap SsoLoginService network calls in IOException catch

Notifications:
- Remove USE_EXACT_ALARM (Play Store policy violation)
- Add RECEIVE_BOOT_COMPLETED permission and BootReceiver to reschedule alarms
- Mask notification ID to positive range to reduce hashCode collisions

UI & State:
- Add SettingsViewModel.refreshLoginState() for stale login state
- Guard CalendarViewModel Moodle event removal with emptiness check
- Pass CoroutineScope to TimeSliderViewModel via rememberCoroutineScope
- Fix LibraryViewModel auto-refresh to retry on failure
- Fix accentColor bit-masking for 32-bit ARGB values
- Fix AppConstants period B/C overlap at 20:15

Data:
- Cache Course.schedule parsing with shared Gson instance
- Fix DataCache date format timezone (literal 'Z' -> Z pattern)
- Add max iteration guard (365) to ICS multi-day event loop
- Add @Singleton to all DAO providers in AppModule
- Prefer moodleUrl for Assignment deep link when available

+++ app/src/main/java/org/ntust/app/tigerduck/network/MoodleService.kt
@@ -57,7 +57,7 @@ class MoodleService @Inject constructor(
                 set(Calendar.SECOND, 0)
                 set(Calendar.MILLISECOND, 0)
             }
-            val startOfToday = (cal.timeInMillis / 1000).toInt()
+            val startOfToday = cal.timeInMillis / 1000
 
             val apiUrl = moodleApiTemplate.format(sesskey)
             val payload = gson.toJson(listOf(MoodleCalendarRequest.upcoming(startOfToday)))
@@ -97,7 +97,7 @@ class MoodleService @Inject constructor(
                     courseNo = courseNo,
                     courseName = courseName,
                     title = event.activityname ?: event.name,
-                    dueDate = Date(event.timestart.toLong() * 1000),
+                    dueDate = Date(event.timestart * 1000),
                     isCompleted = false,
                     moodleUrl = event.url
                 )
